+++ src/lib.rs
@@ -10,7 +10,7 @@ pub mod message;
 pub use crate::class::{MidiClass, MidiReadError};
 pub use crate::packet::cable_number::CableNumber;
 pub use crate::packet::reader::UsbMidiPacketReader;
-pub use crate::packet::{MidiPacketParsingError, UsbMidiEventPacket};
+pub use crate::packet::{UsbMidiEventPacketError, UsbMidiEventPacket};
 
 #[cfg(feature = "message-types")]
 pub use crate::message::Message;